Improve expansion & heuristics
- Added greedy approaches 1 & 2
- Play again
- Repeat simulation
- Added weight to approach 2

+++ MKAgent/UCT.java
@@ -6,7 +6,7 @@ public class UCT {
 
     public static double getUCTValue(int noOfTotalVisits, int totalScore, int noOfVisits) {
         if (noOfVisits == 0)
-            return Integer.MAX_VALUE;
+            return Double.MAX_VALUE;
 
         return (totalScore / (double)noOfVisits + Math.sqrt(2 * Math.log(noOfTotalVisits) / (double)noOfVisits));
     }
